kyq17-master/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-disqus==0.5
django-embed-video==1.2.0
gunicorn==19.9.0
idna==2.8
mock==3.0.5
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
